update brand logo pressed

+++ SneakerStore/app/src/main/java/com/example/sneakerstore/OrderActivity.java
@@ -88,7 +88,7 @@ public class OrderActivity extends AppCompatActivity{
         paymentBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                userOder = new Order(MainActivity.user.getId(), totalProductPrice + Integer.parseInt(costText.getText().toString().split(" ")[0]), receiveOption, (googlePayBox.isChecked()) ? 1 : 0, (shippingBox.isChecked()) ? placeView.getText().toString() : "");
+                userOder = new Order(MainActivity.user.getId(), totalProductPrice + Double.parseDouble(costText.getText().toString().split(" ")[0]), receiveOption, (googlePayBox.isChecked()) ? 1 : 0, (shippingBox.isChecked()) ? placeView.getText().toString() : "");
                 if (googlePayBox.isChecked()) {
                     Intent intent = new Intent(OrderActivity.this, OrderSuccessActivity.class);
                     startActivityForResult(intent, 30);
